requirements.txt
discord.py
python-dotenv
matplotlib 
pandas 
pillow
wordcloud
pytz
fish-audio-sdk